dispersion_signal/requirements.txt
supabase==2.0.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pandas==2.1.0
